WIP compilation fixes
